create api for get company by id

+++ spring-angular/src/main/java/com/example/springangularapp/repository/CompanyRepository.java
@@ -5,7 +5,7 @@ import com.example.springangularapp.entity.Company;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface CompanyRepository extends CrudRepository<Company, Long> {
+public interface CompanyRepository extends CrudRepository<Company, Integer> {
     Company findByEmail(String email);
 
 }
